* 'develop' of http://github.com/pydio/pydio-core:
  Ignore compiled css
  Remove css compilation from jspack
  Fix Connexion class: commited ES6 in ES5 context.
  Add check hook on pluploader. Manual merge and close #957
  Exception namespace
  Manual merge of PR 1202 - Close #1202
  Share info panel : fix copy bytton style. Share: avoid using null repoId. Server catchError: do not use error code as request status !
  Change cache invalidation method, by timestamping cached data and checking a given key against other data stamps. To be tested extensively.
  Simplify S3 as we embed the sdk, no need to handle v2 / v3 case anymore, VHOST_UNSUPPORTED not required either.

+++ core/src/plugins/action.share/src/Store/ShareStore.php
@@ -440,7 +440,7 @@ class ShareStore {
                 $owner = $repo->getOwner();
                 $this->testUserCanEditShare($repo->getOwner(), $repo->options);
             }
-            if(!$keepRepository){
+            if($repoId !== null && !$keepRepository){
                 $res = RepositoryService::deleteRepository($repoId);
                 if ($res == -1) {
                     throw new \Exception($mess[427]);
